+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# 0.1.0+2
+
+- docs: fix README image
+- feat: add faculty of renewable and natural resources and departments
+- feat: add faculty of law and departments
+- feat: add faculty of veterinary medicine and departments
+- chore: add missing departments for exixting faculties
+- feat: change "basic medical sciences" to "college of medicine"
+
 # 0.1.0+1
 
 - 🎉: Create initial release for `unibadan_template` brick.